Extract common utilities for simple and complex polygon generation
